Fixed bug with GROW, SCAN and WIPE emerged when using one module only.

+++ src/MD_Parola_Scan.cpp
@@ -59,7 +59,7 @@ void MD_PZone::effectHScan(bool bIn, bool bBlank)
 
       // blank out the part of the display we don't need
       FSMPRINT("Scan col ", _nextPos);
-      for (uint8_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
+      for (int16_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
       {
         if ((!bBlank && (i != _nextPos)) || (bBlank && (i == _nextPos)))
           _MX->setColumn(i, EMPTY_BAR);
@@ -92,7 +92,7 @@ void MD_PZone::effectHScan(bool bIn, bool bBlank)
 
       // blank out the part of the display we don't need
       FSMPRINT(" Scan col ", _nextPos);
-      for (uint8_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
+      for (int16_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
       {
         if ((!bBlank && (i != _nextPos)) || (bBlank && (i == _nextPos)))
           _MX->setColumn(i, EMPTY_BAR);
@@ -148,7 +148,7 @@ void MD_PZone::effectVScan(bool bIn, bool bBlank)
       // blank out the part of the display we don't need
       FSMPRINT("Keep bit ", _nextPos);
       maskCol = (1 << _nextPos);
-      for (uint8_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
+      for (int16_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
       {
         uint8_t	c = DATA_BAR(_MX->getColumn(i) & (bBlank ? ~maskCol : maskCol));
 
@@ -186,7 +186,7 @@ void MD_PZone::effectVScan(bool bIn, bool bBlank)
       FSMPRINT(" Keep bit ", _nextPos);
       if (_nextPos >= 0)
         maskCol = 1 << _nextPos;
-      for (uint8_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
+      for (int16_t i=_startPos; i != _endPos+_posOffset; i += _posOffset)
       {
         uint8_t	c = DATA_BAR(_MX->getColumn(i) & (bBlank ? ~maskCol : maskCol));
